Make some LDAP parameters conditional around FreeRADIUS v3.1.x
Impacted params:
- connect_timeout
- session_tracking
- use_referral_credentials

Only set each of these params in the config if:
- it is intentionally set in the params, or
- we are on FR3.1.x, in which case set the default of 3.0

If they are set in the params, and we detect we are NOT on FR3.1.x, then warn

Additionally, move the LDAP module tests to the module directory
